% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_bands}
\alias{coverage_get_bands}
\title{Get Bands}
\usage{
coverage_get_bands(desc_url, coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character]}

\item{coverage}{Name of a coverage [character]}
}
\description{
Get the available bands of one coverage
}
