% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_bounding_box}
\alias{coverage_get_bounding_box}
\title{Get Bounding Box}
\usage{
coverage_get_bounding_box(coverage, url = NULL)
}
\arguments{
\item{coverage}{character; Name of a coverage}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
Get a bounding Box of a WCS, WCPS coverage
}
