% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_coordinate_reference}
\alias{coverage_get_coordinate_reference}
\title{Get EPSG identifier}
\usage{
coverage_get_coordinate_reference(desc_url = NULL, coord_url = NULL,
  coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character]}

\item{coord_url}{character; Url for retrieving EPSG code [character]}

\item{coverage}{Name of a coverage [character]}
}
\description{
Function to extract the EPSG identifier from a WCS WCPS coverage
}
