% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_coordsys}
\alias{coverage_get_coordsys}
\title{Get coordinate system}
\usage{
coverage_get_coordsys(desc_url = NULL, coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character].
This URL can be built with the *createWCS_URLs* funtion}

\item{coverage}{Name of a coverage [character]}
}
\description{
Get the coordinate system of a coverage
}
