% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_number_timestamps}
\alias{coverage_get_number_timestamps}
\title{Get Number Timestamps}
\usage{
coverage_get_number_timestamps(desc_url = NULL, coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character]
This URL can be built with the *createWCS_URLs* funtion}

\item{coverage}{Name of a coverage [character]}
}
\description{
Get the number of the available timestamps of WCS, WCPS coverage
}
