% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_resolution}
\alias{coverage_get_resolution}
\title{Get Resolution}
\usage{
coverage_get_resolution(coverage, url = NULL)
}
\arguments{
\item{coverage}{character; Name of a coverage}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
Get the resolution of one WCS, WCPS coverage
}
