% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_timestamps}
\alias{coverage_get_timestamps}
\title{Get Timestamps}
\usage{
coverage_get_timestamps(coverage, url = NULL, timetype = FALSE,
  verbose = FALSE)
}
\arguments{
\item{coverage}{character; Name of a coverage}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}

\item{timetype}{boolean; return the type of storage of time dimension in coverage (regular, irregular, none)?}

\item{verbose}{boolean; Want to return the processed coverage?}
}
\description{
Get the available timestamps of WCS coverages
}
