% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{createWCS_URLs}
\alias{createWCS_URLs}
\title{WCS URL creator}
\usage{
createWCS_URLs(type, url = NULL)
}
\arguments{
\item{type}{character; Mandatory input for the URL that should be created with the function.
For now the types "Meta", "Query","Capabilities" can be created.}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function provides the possibility to construct the url necessary to address different layers of
Metadata. Every Metadata Information in  the WCS OGS format has to be addressed with a differen URL Query.
}
