% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cube Metadata Assay}
\alias{Cube Metadata Assay}
\alias{metadata_assay}
\title{Cube Metadata Assay}
\format{
A assay metadata with 18 rows and 5 variables:
\describe{
\item{element_id}{assay id}
\item{element_label}{name of assay}
\item{instance_id}{assay instance id}
\item{date_created}{date created}
\item{property value}{assay property value}
}
}
\usage{
metadata_assay
}
\description{
Metadata information for assay
}
\keyword{datasets}
