% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cube Metadata Drug_Treatment}
\alias{Cube Metadata Drug_Treatment}
\alias{drug_treatment}
\title{Cube Metadata Drug_Treatment}
\format{
A Disease metadata with 3 variables:
\describe{
\item{accession_id}{accession id}
\item{chebi_name}{Ch E B name }
\item{chebi_id}{Ch E B Id}
\item{short_name}{Short Name}
}
}
\usage{
drug_treatment
}
\description{
Metadata information for Drug_Treatment
}
\keyword{datasets}
