% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_apply_.R
\name{cube_}
\alias{cube_}
\title{Create a data cube and store it in a list}
\usage{
cube_(.data, .vars, sep = ".", fun_list = list(n = ~nrow(.data)), ...)
}
\arguments{
\item{.data}{a data frame.}

\item{.vars}{character vector with the names of the columns for which the 
data frame will be divided.}

\item{sep}{a character string to separate the terms.}

\item{fun_list}{a named list of formulas with the desired function calls to
be evalutated on the data frame. The default value (\code{list(n = ~nrow(.data))})
returns the number of rows (observations) and stores it in a column named \code{n}.}

\item{...}{additional optional arguments passed to \code{\link{res2df}} 
(\code{estimator_name} and \code{value}).}
}
\value{
a list whose elements are data frame containing the estimates for 
each combination of variables.
}
\description{
Create a data cube and store it in a list
}
\examples{
data(invented_wages)
str(invented_wages)

# First create a list of function calls (with formulas) and
# a vector with names of the desired variables
fun <- list(m1 = ~weighted.mean(wage, sample_weights), s = ~sum(sample_weights))
vars <- c("gender", "sector")

# And then create the data cube, with the estimates of all the combinations
# of variables
cube_(invented_wages, .vars = vars, fun_list = fun)
cube_(invented_wages, .vars = vars, fun_list = fun, 
      estimator_name = "statistica", value = "stima")

}
