% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_apply_.R
\name{cube_df_}
\alias{cube_df_}
\title{Create a data cube and store it in a data frame}
\usage{
cube_df_(
  .data,
  .vars,
  sep = ".",
  fun_list = list(n = ~nrow(.data)),
  .total = "Totale",
  ...
)
}
\arguments{
\item{.data}{a data frame.}

\item{.vars}{character vector with the names of the columns for which the 
data frame will be divided.}

\item{sep}{a character string to separate the terms.}

\item{fun_list}{a named list of formulas with the desired function calls to
be evalutated on the data frame. The default value (\code{list(n = ~nrow(.data))})
returns the number of rows (observations) and stores it in a column named \code{n}.}

\item{.total}{character string with the value to give to the "total" combinations
(default: "Totale")}

\item{...}{additional optional arguments passed to \code{\link{res2df}} 
(\code{estimator_name} and \code{value}).}
}
\value{
data frame containing the estimates for each combination of variables.
}
\description{
Create a data cube and store it in a data frame
}
\examples{
data(invented_wages)
str(invented_wages)

# First create a list of function calls (with formulas) and
# a vector with names of the desired variables
fun <- list(m1 = ~weighted.mean(wage, sample_weights), s = ~sum(sample_weights))
vars <- c("gender", "sector")

# And then create the data cube, with the estimates of all the combinations
# of variables
cube_df_(invented_wages, .vars = vars, fun_list = fun)
cube_df_(invented_wages, .vars = vars, fun_list = fun, 
         estimator_name = "statistica", value = "stima")

}
