% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_function.R
\name{cube_function}
\alias{cube_function}
\title{A Cube Function}
\usage{
cube_function(dataset_id = "snp_grid")
}
\arguments{
\item{dataset_id}{dataset to access.}
}
\description{
This function allows you to access the cube data.
}
\examples{
cube_function()
}
\keyword{cube}
