% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_apply_.R
\name{cube_to_df_}
\alias{cube_to_df_}
\title{Put the results of a data cube in a data frame}
\usage{
cube_to_df_(.cube, .total = "Totale")
}
\arguments{
\item{.cube}{a data cube. This can be the output of \code{\link{cube_}} or
\code{\link{cube_vars_}}}

\item{.total}{character string with the value to give to the "total" 
combinations (default: "Totale")}
}
\description{
Put the results of a data cube in a data frame
}
