% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuber-package.R
\docType{package}
\name{cuber-package}
\alias{cuber}
\alias{cuber-package}
\title{cuber: Tools to Create Data Cubes}
\description{
Tools to create data cubes. A data cube is a summary of a dataset with estimated statistics for all the combinations of the desired variables.
}
\keyword{internal}
