% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubeR.R
\name{Cuber.R}
\alias{Cuber.R}
\title{Sum of Cubed Numbers}
\usage{
Cuber.R(x, y)
}
\arguments{
\item{x}{is a number}

\item{y}{is a number}
}
\value{
The sum of x^3 and y^3
}
\description{
Sum of Cubed Numbers
}
\examples{
sum.cubed(3,4)
sum.cubed(52, 96)
}
