% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubeR.R
\name{cubeR}
\alias{cubeR}
\title{A CubeR Function}
\usage{
cubeR(dataset_id = "jaxda00234")
}
\arguments{
\item{dataset_id}{dataset to access.}
}
\description{
This is basic function to retrieve cube data
}
\examples{
cubeR("jaxda00234")
}
\keyword{cubeR}
