% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{findCoverages}
\alias{findCoverages}
\title{Find Capabilities}
\usage{
findCoverages(str, url = url)
}
\arguments{
\item{str}{character; string to search for among coverages}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function searches for available Coverages within the
}
