% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoded_pixel_buffer.R
\name{geocoded_pixel_buffer}
\alias{geocoded_pixel_buffer}
\title{Calculate Pixel Buffer History}
\usage{
geocoded_pixel_buffer(coverage, coords, band, buffer, date = NULL,
  filename = "request", url = NULL, plot = F)
}
\arguments{
\item{coverage}{character; name of the coverage}

\item{coords}{character; coordinates of the location of interest}

\item{band}{character; coverage band}

\item{buffer}{numeric; A buffer Zone around the coordinates of the Poi}

\item{date}{character; date range in format (Ymd)}

\item{filename}{character; filename for the NETCDF Output. If none is provided the Object will be deleted after temporal storage in the tempdir}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}

\item{plot}{boolean; do you want a generic plot to be returned?}
}
\description{
Calculate pixel buffer history of a coverage band. This functionality is limited to Netcdf files
}
