% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_data.R
\name{get_coverage}
\alias{get_coverage}
\title{GetCoverage}
\usage{
get_coverage(
  coverage,
  DATA,
  FORMAT,
  SUBSET_E = NULL,
  SUBSET_N = NULL,
  BAND = NULL,
  CRS_Extension = NULL,
  filename = NULL,
  others_opt = NULL
)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{DATA}{an available timestamp [character]}

\item{FORMAT}{output_format. Output format (mime type) of the product}

\item{BAND}{(RangeSubset) coverage Nameband to calculate raster. If param BAND not set (=NULL) the band number 1 will be automatically selected. [character]}

\item{CRS_Extension}{This parameter defines the output crs. By default is used the coverages CRS. [character] Eg. http://localhost:8080/def/crs/EPSG/0/32632}

\item{filename}{If the raster image should be saved please digit a path and a filename. [character]}

\item{others_opt}{Other options/WCS GetCoverage parameters. Eg. clipping extension: clip=POLYGON((1520000 5030000,1540000 5030000, 1540000 5060000, 1520000 5060000, 1520000 5030000))}

\item{slice_E}{image slicing coordinates in x-direction [character]  Eg. E(436000, 550000)}

\item{slice_N}{image slicing coordinates in y-direction [character]  Eg. N(4918000,5166000)}
}
\description{
This function provides the possibility to interact directly with the data cubes. It gives the option to write the images to memory or on the hard drive for further computation.
}
