% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_person_data.R
\name{get_person_data}
\alias{get_person_data}
\title{get_person_data}
\usage{
get_person_data(id, export_csv = FALSE, directory = NULL, file_name)
}
\arguments{
\item{id}{ID of the person to be searched.}
}
\description{
Gets person data from WCA website using the ID provided.
}
