% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record.R
\name{get_record}
\alias{get_record}
\title{get_record()}
\usage{
get_record(event, region = "World")
}
\arguments{
\item{event}{Event of the record to be searched for. Note that we use
the short version of the event name, e.g. 3x3 is 333, Square-1 is sq1, and so on}

\item{region}{Region of the record to be search for (World on default)}
}
\description{
Gets the record of the chosen event and region
}
