% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regulation.R
\name{get_regulation}
\alias{get_regulation}
\title{get_regulation}
\usage{
get_regulation(id)
}
\arguments{
\item{id}{Number or ID of the regulation to be searched for.}
}
\description{
Gets the chosen regulation from the WCA website.
}
