% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{getCapabilities}
\alias{getCapabilities}
\title{Returns the Capabilities}
\usage{
getCapabilities(url = NULL)
}
\arguments{
\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function Returns the Capabilities of a DataCube containing all coverages available
}
