% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{getCapability}
\alias{getCapability}
\title{Returns the Capabilities}
\usage{
getCapability(url = NULL)
}
\arguments{
\item{url}{This central URL Leads to the 'ows' page of the Datacube}
}
\description{
This function Returns the Capabilities of a DataCube containing all coverages available
}
