% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{getCoverage}
\alias{getCoverage}
\title{List Coverages}
\usage{
getCoverage(url = NULL)
}
\arguments{
\item{url}{This central URL Leads to the 'ows' page of the Datacube}
}
\description{
This function Lists all Coverage inluced in the Datacube
}
