% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{getFormats}
\alias{getFormats}
\title{Rasdaman support formats}
\usage{
getFormats(url = NULL)
}
\arguments{
\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function returns the formats supported by the Rasdaman Version
}
