% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Get coverage metadata}
\usage{
getMetadata(coverage, Metacols = NULL, url = NULL)
}
\arguments{
\item{coverage}{character; Name of a coverage}

\item{Metacols}{tibble; 2 column tibble with all available columns and their dimension
If NULL the standard SAO columns will be used}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function retieves the actual Metadata from the
containing both values and attributes. Other *coverage_get* functions are built on top of this function
}
\note{
used in the getMetadata function
}
