% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{getProvider}
\alias{getProvider}
\title{Returns the Service Provider of the Rasdaman Environment}
\usage{
getProvider(url = NULL)
}
\arguments{
\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function returns the contact information to the Rasdaman service provider
}
