% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_data.R
\name{image_from_coverage}
\alias{image_from_coverage}
\title{Get a raster from coverage}
\usage{
image_from_coverage(coverage, coord_sys, slice_E, slice_N, date,
  ref_Id = NULL, res_eff = NULL, format = "TIFF", bands = NULL,
  pixel_url = NULL, filename = NULL)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coord_sys}{coordinate system [character]}

\item{slice_E}{image slicing coordinates in x-direction [character]}

\item{slice_N}{image slicing coordinates in y-direction [character]}

\item{date}{an available timestamp [character]}

\item{ref_Id}{EPSG code of the coordinate system [character]}

\item{res_eff}{factor to scale raster resolution [numeric]}

\item{format}{image format in WCPS query [character]}

\item{bands}{coverage bands to calculate raster. Can contain one or more bands from the same coverage [character]}

\item{pixel_url}{Web Coverage Service (WCS) for processing the query. This URL can be built with the *createWCS_URLs* function. [character]}

\item{filename}{If the raster image should be saved please digit a path and a filename. [character]}
}
\description{
This function provides the possibility to interact directly with the data cubes. It gives the option to write the images to
memory or on the hard drive for further computation.
}
