% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_difference.R
\name{norm_diff_pixel}
\alias{norm_diff_pixel}
\title{Calculate Normalized Difference Pixel History}
\usage{
norm_diff_pixel(coverage, coords, band1, band2, date = NULL,
  url = NULL, format = "text/csv")
}
\arguments{
\item{coverage}{character; name of the coverage}

\item{coords}{character; coordinates of the location of interest}

\item{band1}{character; coverage band}

\item{band2}{character; coverage band}

\item{date}{character(s); date with one or date range range with two dates. If NULL all the available time range is returned.}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}

\item{format}{character; Output format. Not all available are implemented for each datacube. Please see `getFormat` for all the available formats}
}
\description{
Calculate normalized difference pixel history between band1 and band2
}
