% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_difference.R
\name{norm_diff_raster}
\alias{norm_diff_raster}
\title{Calculate Normalized Difference Raster Layer}
\usage{
norm_diff_raster(coverage, slice_E, slice_N, date, band1, band2,
  res_eff = 1, format = "image/tiff", url = NULL)
}
\arguments{
\item{coverage}{character; name of the coverage}

\item{slice_E}{character; image slicing coordinates in x-direction}

\item{slice_N}{character; image slicing coordinates in y-direction}

\item{date}{character; an available timestamp}

\item{band1}{character; coverage band}

\item{band2}{character; coverage band}

\item{res_eff}{numeric; factor to scale raster resolution}

\item{format}{character; Output format. Not all available are implemented for each datacube. Please see `getFormat` for all the available formats}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
Calculate a raster layer of normalized difference between band_1 and band_2
}
