% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{parseCoverages}
\alias{parseCoverages}
\title{Parse the whole Rasdaman Environment}
\usage{
parseCoverages(url = NULL)
}
\arguments{
\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
Scrape the data in the whole Rasdaman Environment
}
