% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{parseMetadata}
\alias{parseMetadata}
\title{Parse coverage metadata}
\usage{
parseMetadata(coverage, returnxml = F, url = NULL)
}
\arguments{
\item{coverage}{character; Name of a coverage}

\item{returnxml}{You want the Coverage Metadata XML return?}

\item{url}{character; Web Coverage Service (WCS) DescribeCoverage url
If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function parses the metadata of a coverage and returns a tibble
containing both values and attributes. Other *coverage_get* functions are built on top of this function
}
\note{
used in the getMetadata function
}
