% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_history.R
\name{pixel_history}
\alias{pixel_history}
\title{Pixel History}
\usage{
pixel_history(coverage, coord_sys, bands, coords, date = NULL)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coord_sys}{coordinate system [character]}

\item{bands}{coverage bands to calculate and visualize pixel history [character]}

\item{coords}{coordinates of the location of interest [character]}

\item{date}{date range in format (Ymd) [character]}
}
\description{
Pixel history of "bands" (coverage bands) for a given point determined "coords"
}
