% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_apply_.R
\name{res2df}
\alias{res2df}
\title{Put the results of 'split_apply_' in a data frame}
\usage{
res2df(.nested_list, sep = ".", estimator_name = "estimator",
  value = "value")
}
\arguments{
\item{.nested_list}{a nested list, the output of \code{split_apply_}.}

\item{sep}{a character string to separate the terms.}

\item{estimator_name}{character string with the name to be given to the
column with the type of the estimated statistics (default: "estimator").}

\item{value}{character string with the name of the column that will contain 
the values of the estimated statistics (default: "value").}
}
\value{
a data frame
}
\description{
Put the results of 'split_apply_' in a data frame
}
\examples{
data(invented_wages)
str(invented_wages)

# First create a list of function calls (with formulas)
fun <- list(m1 = ~weighted.mean(wage, sample_weights), s = ~sum(sample_weights))

# And then apply them to subsets of data defined by the combinations of variables
tmp <- split_apply_(invented_wages, .vars = c("gender", "sector"), fun_list = fun)
str(tmp)     # a list

res2df(tmp)  # a data frame

}
