% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{setMetaCols}
\alias{setMetaCols}
\title{Set possible coverage Metadata axis names}
\usage{
setMetaCols(addcols = NULL, keepsao = F)
}
\arguments{
\item{addcols}{Add dimension columns and the respective name. If NULL the standard SAO columns are used}

\item{keepsao}{boolean; You want to keep the ones from SAO adjacent to the ones you defined?}
}
\description{
Scrape the data in the whole Rasdaman Environment
}
