% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNP Grid Data}
\alias{SNP Grid Data}
\alias{snp_data}
\title{SNP Grid Data}
\format{
A SNP data frame with 1000 rows and 10 variables:
\describe{
\item{chr}{chromosome number, eg: 6}
\item{start_position}{start position}
\item{end_position}{end position}
\item{rs}{rs}
\item{strain_id}{strain id}
\item{strain_name}{strain name}
\item{genotyp}{genotype}
\item{imputed_flag}{imputed flag}
\item{percent_impluted}{percent impluted}
\item{percent_missing}{percent missing}
}
}
\source{
\url{https://www.github.com/TheJacksonLaboratory/cube_r}
}
\usage{
snp_data
}
\description{
A sample data set for SNP Grid Data.
}
\keyword{datasets}
