% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_data.R
\name{WPCS_query}
\alias{WPCS_query}
\title{Get WPCS query result}
\usage{
WPCS_query(
  proper_query = NULL,
  formato = NULL,
  filename = NULL,
  query_url = NULL
)
}
\arguments{
\item{proper_query}{The proper WPCS query for the selected coverage, in WPCS query language (eg. 'for c in (rh_ana) return encode(c[E(515200),N(5037430), ansi("2020-10-01")], "text/csv" )'. The script takes care of encoding it}

\item{formato}{output format for the WCPS query result. [character]
The format coul be image format (image/png, image/tiff, image/jpeg, image/bmp) or text/csv format (text/csv)}

\item{filename}{downloadfile name (all path)}

\item{query_url}{Web Coverage Service (WCS) for processing the query. This URL can be built with the *createWCS_URLs* function. [character]}
}
\description{
This function provides the possibility to interact directly with the data cubes. It gives the option to execute a WPCS query in txt/csv format
}
