% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_bounding_box}
\alias{coverage_get_bounding_box}
\title{Get Bounding Box}
\usage{
coverage_get_bounding_box(desc_url = NULL, coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character]
This URL can be built with the *createWCS_URLs* funtion}

\item{coverage}{Name of a coverage [character]}
}
\description{
Get a bounding Box of a WCS, WCPS coverage
}
