% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_coordinate_reference}
\alias{coverage_get_coordinate_reference}
\title{Get EPSG identifier}
\usage{
coverage_get_coordinate_reference(coverage, url = NULL)
}
\arguments{
\item{coverage}{character; Name of a coverage}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
Function to extract the EPSG identifier from a WCS WCPS coverage
}
