% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_temporal_extent}
\alias{coverage_get_temporal_extent}
\title{Get Temporal Extent}
\usage{
coverage_get_temporal_extent(desc_url = NULL, coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character]
This URL can be built with the *createWCS_URLs* funtion}

\item{coverage}{Name of a coverage [character]}
}
\description{
Get the temporal extent from a WCS WCPS Coverage
}
