% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{coverage_get_timestamps}
\alias{coverage_get_timestamps}
\title{Get Timestamps}
\usage{
coverage_get_timestamps(desc_url, coverage)
}
\arguments{
\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character]}

\item{coverage}{Name of a coverage [character]}
}
\description{
Get coverage's available timestamps (images)
}
