% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_metadata.R
\name{createWCS_URLs}
\alias{createWCS_URLs}
\title{WCS URL creator}
\usage{
createWCS_URLs(url = NULL, type)
}
\arguments{
\item{url}{This central URL Leads to the 'ows' page of the Datacube. If empy the Standard Rasdaman page is used.}

\item{type}{characer; Mandatory input for the URL that should be created with the function.
For now the types "Meta", "Pixel","Tiff","Coords","Time" can be created.}
}
\description{
This function provides the possibility to construct the url necessary to address different layers of
Metadata. Every Metadata Information in  the WCS OGS format has to be addressed with a differen URL Query.
}
