% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cube Metadata Assay}
\alias{Cube Metadata Assay}
\alias{assay}
\title{Cube Metadata Assay}
\format{
A assay metadata with 5 variables:
\describe{
\item{accession_id}{accession id}
\item{subclass}{Sub class}
\item{target}{Target}
\item{seqclass}{Seq class}
\item{short}{Short}
}
}
\usage{
assay
}
\description{
Metadata information for assay
}
\keyword{datasets}
