% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoded_pixel_buffer.R
\name{geocoded_pixel_buffer}
\alias{geocoded_pixel_buffer}
\title{Calculate Pixel Buffer History}
\usage{
geocoded_pixel_buffer(coverage, coord_sys, coords, band, buffer, date = NULL,
  pixel_url = NULL)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coord_sys}{coordinate system [character]}

\item{coords}{coordinates of the location of interest [character]}

\item{band}{coverage band [character]}

\item{date}{date range in format (Ymd) [character]}

\item{pixel_url}{Web Coverage Service (WCS) for processing the query [character].
This URL can be built with the *createWCS_URLs* function}
}
\description{
Calculate pixel buffer history of a coverage band
}
