% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_data.R
\name{get_comp_data}
\alias{get_comp_data}
\title{get_comp_data}
\usage{
get_comp_data(id, export_csv = FALSE, directory = NULL, file_name)
}
\arguments{
\item{id}{ID of the competition to be searched for.}

\item{export_csv}{Whether or not the data should be exported. Set to false by default.}

\item{directory}{Directory where the .csv file will be saved.}

\item{file_name}{File name of the .csv file.}
}
\description{
Gets competition data from WCA using the ID provided.
}
