% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_data.R
\name{image_from_coverage}
\alias{image_from_coverage}
\title{Get a raster from the coverage}
\usage{
image_from_coverage(coverage, coord_sys, slice_E, slice_N, date, ref_Id,
  res_eff, format, bands)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coord_sys}{coordinate system [character]}

\item{slice_E}{image slicing coordinates in x-direction [character]}

\item{slice_N}{image slicing coordinates in y-direction [character]}

\item{date}{an available timestamp [character]}

\item{ref_Id}{EPSG code of the coordinate system [character]}

\item{res_eff}{factor to scale raster resolution [numeric]}

\item{format}{image format in WCPS query [character]}

\item{bands}{coverage bands to calculate raster [character]}
}
\description{
Calculate raster layer(s) of the coverage
}
