% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_data.R
\name{image_from_coverage}
\alias{image_from_coverage}
\title{Get a raster from coverage}
\usage{
image_from_coverage(coverage, slice_E, slice_N, date, res_eff = 1,
  format = "image/tiff", bands = NULL, filename = NULL, url = NULL)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{slice_E}{image slicing coordinates in x-direction [character]}

\item{slice_N}{image slicing coordinates in y-direction [character]}

\item{date}{an available timestamp [character]}

\item{res_eff}{factor to scale raster resolution [numeric]}

\item{format}{image format in WCPS query [character]}

\item{bands}{character; coverage bands to calculate raster. Can contain one or more bands from the same coverage.
If NULL all bands are used}

\item{filename}{If the raster image should be saved please digit a path and a filename. [character]}

\item{url}{Web Coverage Service (WCS) DescribeCoverage url [character].
If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}
}
\description{
This function provides the possibility to interact directly with the data cubes. It gives the option to write the images to
memory or on the hard drive for further computation.
}
