% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_difference.R
\name{norm_diff_pixel}
\alias{norm_diff_pixel}
\title{Calculate Normalized Difference Pixel History}
\usage{
norm_diff_pixel(coverage, coords, band1, band2, date = NULL,
  desc_url = NULL, query_url = NULL, plot = TRUE)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coords}{coordinates of the location of interest [character]}

\item{band1}{coverage band [character]}

\item{band2}{coverage band [character]}

\item{date}{date range in format (Ymd) [character]}

\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character].
This URL can be built with the *createWCS_URLs* function}

\item{query_url}{Web Coverage Service (WCS) for processing the query [character].
This URL can be built with the *createWCS_URLs* function}

\item{plot}{handler if a plot is returned or a vector containing timestamp and value}
}
\description{
Calculate normalized difference pixel history between band1 and band2
}
