% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_difference.R
\name{norm_diff_raster}
\alias{norm_diff_raster}
\title{Calculate Normalized Difference Raster Layer}
\usage{
norm_diff_raster(coverage, coord_sys, slice_E, slice_N, date, ref_Id, res_eff,
  format, band_1, band_2)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coord_sys}{coordinate system [character]}

\item{slice_E}{image slicing coordinates in x-direction [character]}

\item{slice_N}{image slicing coordinates in y-direction [character]}

\item{date}{an available timestamp [character]}

\item{ref_Id}{EPSG code of the coordinate system [character]}

\item{res_eff}{factor to scale raster resolution [numeric]}

\item{format}{image format in WCPS query [character]}

\item{band_1}{coverage band [character]}

\item{band_2}{coverage band [character]}
}
\description{
Calculate a raster layer of normalized difference between band_1 and band_2
}
