% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_difference.R
\name{norm_diff_raster}
\alias{norm_diff_raster}
\title{Calculate Normalized Difference Raster Layer}
\usage{
norm_diff_raster(coverage, slice_E, slice_N, date, band1, band2,
  res_eff = 1, format = "TIFF", query_url = NULL)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{slice_E}{image slicing coordinates in x-direction [character]}

\item{slice_N}{image slicing coordinates in y-direction [character]}

\item{date}{an available timestamp [character]}

\item{band1}{coverage band [character]}

\item{band2}{coverage band [character]}

\item{res_eff}{factor to scale raster resolution [numeric]}

\item{format}{image format in WCPS query [character]}

\item{query_url}{Web Coverage Service (WCS) for processing the query [character].
This URL can be built with the *createWCS_URLs* function}
}
\description{
Calculate a raster layer of normalized difference between band_1 and band_2
}
