% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_history.R
\name{pixel_history}
\alias{pixel_history}
\title{Pixel History}
\usage{
pixel_history(coverage, coords = NULL, bands = NULL, date = NULL,
  url = NULL, encoding = "UTF-8")
}
\arguments{
\item{coverage}{character; name of the coverage}

\item{coords}{character; coordinates of the location of interest in the native coverage projection. If NULL the middle coordinate will be taken}

\item{bands}{character; coverage bands to calculate and visualize pixel history}

\item{date}{character; date range in format (Ymd)}

\item{url}{character; Web Coverage Service (WCS) Url. If NULL then it is directing to the SAO homepage ("http://saocompute.eurac.edu/rasdaman/ows")}

\item{encoding}{character; Encoding for the response. "UTF-8" as standard}
}
\description{
Returns a pixel history of one or multiple bands from one coverage.
}
