% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_history.R
\name{pixel_history}
\alias{pixel_history}
\title{Pixel History}
\usage{
pixel_history(coverage, coord_sys, bands, coords, date = NULL,
  desc_url = NULL, query_url = NULL, plot = F)
}
\arguments{
\item{coverage}{name of the coverage [character]}

\item{coord_sys}{coordinate system [character]}

\item{bands}{coverage bands to calculate and visualize pixel history [character]}

\item{coords}{coordinates of the location of interest in c(Lat,Lon)[character]}

\item{date}{date range in format (Ymd) [character]}

\item{desc_url}{Web Coverage Service (WCS) DescribeCoverage url [character].
This URL can be built with the *createWCS_URLs* function}

\item{query_url}{Web Coverage Service (WCS) for processing the query [character].
This URL can be built with the *createWCS_URLs* function}

\item{plot}{Should the data be plotted [boolean].}
}
\description{
Returns a pixel history of one or multiple bands from one coverage.
}
